ai-part/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.6.0
numpy==2.2.1
requests==2.32.3
schedule==1.2.2
beautifulsoup4==4.12.3
openai>=1.50.0
httpx>=0.27.2
pydantic==2.10.3